+++ packages/eas-cli/src/commands/workflow/create.ts
@@ -1,3 +1,4 @@
+import { Flags } from '@oclif/core';
 import chalk from 'chalk';
 import fs from 'fs/promises';
 import fsExtra from 'fs-extra';
@@ -5,23 +6,20 @@ import path from 'path';
 import prompts from 'prompts';
 
 import EasCommand from '../../commandUtils/EasCommand';
-import { EASNonInteractiveFlag } from '../../commandUtils/flags';
+import {
+  WorkflowStarter,
+  WorkflowStarterName,
+  customizeTemplateIfNeededAsync,
+  workflowStarters,
+} from '../../commandUtils/workflow/creation';
+import {
+  logWorkflowValidationErrors,
+  validateWorkflowFileAsync,
+  workflowContentsFromParsedYaml,
+} from '../../commandUtils/workflow/validation';
 import Log from '../../log';
+import { promptAsync } from '../../prompts';
 import { WorkflowFile } from '../../utils/workflowFile';
-
-const DEFAULT_WORKFLOW_NAME = 'workflow.yml';
-const HELLO_WORLD_TEMPLATE = `name: Hello World
-
-on:
-  push:
-    branches: ['*']
-
-jobs:
-  hello_world:
-    steps:
-      - uses: eas/checkout
-      - run: echo "Hello, World"
-`;
 
 export class WorkflowCreate extends EasCommand {
   static override description = 'create a new workflow configuration YAML file';
@@ -35,11 +33,20 @@ export class WorkflowCreate extends EasCommand {
   ];
 
   static override flags = {
-    ...EASNonInteractiveFlag,
+    template: Flags.enum({
+      description: 'Name of the template to use',
+      options: Object.values(WorkflowStarterName),
+    }),
+    'skip-validation': Flags.boolean({
+      description: 'If set, the workflow file will not be validated before being created',
+      default: false,
+    }),
   };
 
   static override contextDefinition = {
+    ...this.ContextOptions.DynamicProjectConfig,
     ...this.ContextOptions.ProjectDir,
+    ...this.ContextOptions.LoggedIn,
   };
 
   async runAsync(): Promise<void> {
@@ -48,42 +55,90 @@ export class WorkflowCreate extends EasCommand {
       flags,
     } = await this.parse(WorkflowCreate);
 
-    const { projectDir } = await this.getContextAsync(WorkflowCreate, {
-      nonInteractive: flags['non-interactive'],
-    });
-
-    let fileName = argFileName;
-
-    if (!fileName) {
-      const response = await prompts({
-        type: 'text',
-        name: 'fileName',
-        message: 'What would you like to name your workflow file?',
-        initial: DEFAULT_WORKFLOW_NAME,
-        validate: value => {
-          try {
-            WorkflowFile.validateYamlExtension(value);
-            return true;
-          } catch (error) {
-            return error instanceof Error ? error.message : 'Invalid file name';
-          }
-        },
+    try {
+      const {
+        getDynamicPrivateProjectConfigAsync,
+        loggedIn: { graphqlClient },
+        projectDir,
+      } = await this.getContextAsync(WorkflowCreate, {
+        nonInteractive: false,
+        withServerSideEnvironment: null,
       });
 
-      if (!response.fileName) {
-        Log.warn('Workflow creation cancelled.');
-        process.exit(0);
+      const { exp: expPossiblyWithoutEasUpdateConfigured, projectId } =
+        await getDynamicPrivateProjectConfigAsync();
+
+      let fileName = argFileName;
+
+      let workflowStarter: WorkflowStarter;
+      if (flags.template) {
+        workflowStarter =
+          workflowStarters.find(template => template.name === flags.template) ??
+          workflowStarters[0];
+      } else {
+        workflowStarter = (
+          await promptAsync({
+            type: 'select',
+            name: 'starter',
+            message: 'Select a workflow template:',
+            choices: workflowStarters.map(starter => ({
+              title: starter.displayName,
+              value: starter,
+            })),
+          })
+        ).starter;
       }
 
-      fileName = response.fileName;
-    }
+      if (!fileName) {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
+        const response = await prompts({
+          type: 'text',
+          name: 'fileName',
+          message: 'What would you like to name your workflow file?',
+          initial: workflowStarter.defaultFileName,
+          validate: value => {
+            try {
+              WorkflowFile.validateYamlExtension(value);
+              return true;
+            } catch (error) {
+              return error instanceof Error ? error.message : 'Invalid file name';
+            }
+          },
+        });
+        if (!response.fileName) {
+          Log.warn('Workflow creation cancelled.');
+          process.exit(0);
+        }
 
-    try {
+        fileName = response.fileName;
+      }
+
+      // Customize the template if needed
+      workflowStarter = await customizeTemplateIfNeededAsync(
+        workflowStarter,
+        projectDir,
+        expPossiblyWithoutEasUpdateConfigured
+      );
+
+      Log.debug(`Creating workflow file ${fileName} from template ${workflowStarter.name}`);
+      const yamlString = [
+        ...workflowStarter.headerLines,
+        workflowContentsFromParsedYaml(workflowStarter.template),
+      ].join('\n');
+
+      if (!flags['skip-validation']) {
+        await validateWorkflowFileAsync(
+          { yamlConfig: yamlString, filePath: fileName },
+          projectDir,
+          graphqlClient,
+          projectId
+        );
+      }
       await this.ensureWorkflowsDirectoryExistsAsync({ projectDir });
-      await this.createWorkflowFileAsync({ fileName, projectDir });
+      await this.createWorkflowFileAsync({ fileName, projectDir, yamlString });
     } catch (error) {
+      logWorkflowValidationErrors(error);
       Log.error('Failed to create workflow file.');
-      throw error;
     }
   }
 
@@ -103,9 +158,11 @@ export class WorkflowCreate extends EasCommand {
   private async createWorkflowFileAsync({
     fileName,
     projectDir,
+    yamlString,
   }: {
     fileName: string;
     projectDir: string;
+    yamlString: string;
   }): Promise<void> {
     WorkflowFile.validateYamlExtension(fileName);
 
@@ -115,7 +172,7 @@ export class WorkflowCreate extends EasCommand {
       throw new Error(`Workflow file already exists: ${filePath}`);
     }
 
-    await fs.writeFile(filePath, HELLO_WORLD_TEMPLATE);
+    await fs.writeFile(filePath, yamlString);
     Log.withTick(`Created ${chalk.bold(filePath)}`);
   }
 }